+++ src/main/java/com/github/zastrixarundell/torambot/commands/toramwebsite/Maintenance.java
@@ -21,7 +21,8 @@ public class Maintenance implements MessageCreateListener
             return;
 
         if (!messageCreateEvent.getMessageContent().toLowerCase().startsWith(Values.getPrefix() + "maintenance"))
-            return;
+            if(!messageCreateEvent.getMessageContent().toLowerCase().startsWith(Values.getPrefix() + "maint"))
+                return;
 
         Runnable runnable = () ->
         {